+++ PrivateFrameworks/iTunesStoreUI.framework/SUScriptDebugDelegate.h
@@ -5,7 +5,7 @@
 @interface SUScriptDebugDelegate : NSObject {
 }
 
-- (void)webView:(id)arg1 exceptionWasRaised:(id)arg2 sourceId:(NSInteger)arg3 line:(NSInteger)arg4 forWebFrame:(id)arg5;
+- (void)webView:(id)arg1 exceptionWasRaised:(id)arg2 sourceId:(long)arg3 line:(NSInteger)arg4 forWebFrame:(id)arg5;
 - (void)webView:(id)arg1 failedToParseSource:(id)arg2 baseLineNumber:(NSUInteger)arg3 fromURL:(id)arg4 withError:(id)arg5 forWebFrame:(id)arg6;
 
 @end
